// src/app/models/plantillaRobot.ts
export class plantillaRobot {
    Key: string;
    NumeroCelular: string;
    TipoDocumento: string;
    NumeroDocumento: string;
    IdTipoActivacion: string;
    CodPlanActual: string;
    CodPlanNuevo: string;
    CodBeneficio: string;
    Corte: string;
    EstadoCliente: string;
    FechaProgramadaActivacion: string;
    EnvioContrato: string;
}